+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.12.0] - 2016-0308
 
 * **API** Replace `setClassLayout` with [bindConstructorLayout][doc:bindConstructorLayout].
 * **API** Replace `objectPrototype` with [makeDestinationObject][doc:makeDestinationObject].
@@ -93,7 +93,7 @@
 
 * Initial release.
 
-[Unreleased]: https://github.com/pabigot/buffer-layout/compare/v0.11.0...next
+[0.12.0]: https://github.com/pabigot/buffer-layout/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/pabigot/buffer-layout/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/pabigot/buffer-layout/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/pabigot/buffer-layout/compare/v0.8.0...v0.9.0